factor out BoardClassBoggler
